Modified constructor and implemented makeEdgesToVertices()
Since TotalColoring will calculate a graph's total coloring by substitu-
ting every edge with a vertex and two additional edges and then running
a vertex coloring on that graph, a secure way to add vertices is needed.
Hence, this class uses a VertexFactory to do just that and the
'vertexClass' constructor parameter to construct one.

Also replaced PropertyGraph with Graph for better applicability.